Switched to using status struct and pointers
Problem
- Need to pass the current status of the device to an event

Solution
- Brew_Status struct
- created pointers for the sensor values for readbility

+++ main/brew.h
@@ -33,7 +33,7 @@ typedef struct
   int brewing;
   int error;
   uint32_t setpoint;
-  uint32_t sensors[3];
+  float sensors[3];
 } brewStatus_t;
 
 void brew_init(void);
